requirements.txt
feedparser==6.0.8
python-dateutil==2.8.2
sgmllib3k==1.0.0
six==1.16.0
